controller and model changes

+++ Models/venues.js
@@ -13,10 +13,10 @@ class Venue {
     {
         let response = {};
         await this.venueModel.create(obj).then((created_record)=>{
-            if(!created_record) response.message = "Creation failure";
+            if(!created_record) response.error = "Creation failure";
             else response = created_record;
         }).catch(err=>{
-            response.message = err.message;
+            response.error = err.message;
         });
         return response;
     }
@@ -31,10 +31,12 @@ class Venue {
     async update(existObj, updatedObj)
     {
         let response = {};
+        if (!this.mongoose.Types.ObjectId.isValid(existObj._id))
+            return {error: "Invalid ID"};
         updatedObj.updated_at = new Date().toISOString();
-        await this.venueModel.findOneAndUpdate(existObj, updatedObj).then((updated_record)=>{
-            if(!updated_record) response.message = "Record not found";
-            else response.message = "Update success";
+        await this.venueModel.findOneAndUpdate(existObj, updatedObj, {new: true}).then((updated_record)=>{
+            if(!updated_record) response.error = "Record not found";
+            else response = updated_record;
         }).catch(err=>{
             response.error = err.message;
         });
@@ -43,12 +45,14 @@ class Venue {
     async delete(obj)
     {
         let response = {};
-        await this.venueModel.findOneAndDelete(obj).then((deleted_record)=>{
-            if(!deleted_record) response.message = "Record not found";
-            else response.message = "Delete success";
-        }).catch(err=>{
-            response.error = err.message;
-        });
+        if (this.mongoose.Types.ObjectId.isValid(obj._id))
+            await this.venueModel.findOneAndDelete(obj).then((deleted_record)=>{
+                if(!deleted_record) response.error = "Record not found";
+                else response.message = deleted_record;
+            }).catch(err=>{
+                response.error = err.message;
+            });
+        else response.error = "Invalid ID"
         return response;
     }
 }
